backend/requirements.txt
fastapi==0.118.0
httpx==0.28.1
pydantic[email]==2.11.1
python-dotenv==1.1.1
uvicorn[standard]==0.37.0
